criado estrutura de SOLID para o service product e controller Product
