+++ src/services/product/deleteProductService.ts
@@ -1,0 +1,14 @@
+import { Product } from '../../models/product';
+
+export const deleteProductService = async (id: string): Promise<Product | null> => {
+  try {
+    const product = await Product.findByPk(id);
+    if (product) {
+      await product.destroy();
+      return product;
+    }
+    return null;
+  } catch (error: any) {
+    throw new Error(`Erro ao deletar produto: ${error.message}`);
+  }
+};
